Defines.h: add detection of kernel mode build environments. One of ZYAN_KERNEL or ZYAN_USER will be defined depending on the outcome.
This also implies UEFI detection (because UEFI implies kernel mode), which has been added in a somewhat brittle fashion. Because there are no standardized defines in the UEFI specification that are guaranteed to be present, this detection relies on prior inclusion of Uefi.h (EDK2) or efi.h (gnu-efi), which are both nearly impossible to avoid. However, manually defining ZYAN_UEFI project-wide is still preferable because it avoids this header inclusion order issue.
(Q: Why is the check for UEFI now the very first one? A: Because otherwise ZYAN_WINDOWS might be detected instead, due to MSVC's insistence on defining _WIN32 no matter what)