Add back #ifdef guards to "API" headers/sources for no-libc mode
This essentially reverts ffcc366 (minus CMakeLists.txt) in order to make the "non-compilation" of these files independent of CMake. The #ifndef has been shuffled around somewhat in the various headers so that the intended effect of ffcc366 (silencing compiler warnings about empty TUs) is preserved.

This fixes compilation errors in these TUs when including Zycore in no-libc mode to projects that do not use CMake.