test: make tests runnable with ctest
This helps with automated testing, as tests can be ran with `ctest` or
by "building" the target `test`

I also changed the way the C++ standard version is handled; the tests
require C++17, and this was previously accomplished by manually passing
the `-std=c++17` flag, while the "CMake way" of doing this is to set the
`CXX_STANDARD` property. Unfortunately the value `17` of the property
was only added in CMake 3.8, so I bumped the minimum required version.
I've also set the `CXX_EXTENSIONS` and `CXX_STANDARD_REQUIRED` to false
and true, respectively, to ensure that `-std=c++17` gets passed instead
of `-std=gnu++17` on GCC-like compilers and that CMake doesn't fallback
to older standards versions when the requested standard is not available