Add the ZYAN_DIV64 macro
Add a macro for dividing 64bit numbers.

With 32bit linux targets, GCC generates a call to __udivdi3 when
dividing 64bit integers, even when compiling in freestanding mode.
However, the __udivdi3 implementation resides in libgcc which is not
linked-in when building kernel code.

Besides, the Linux kernel voluntarily doesn't implement __udivdi3 to
force developers to use the do_div macro instead [1].

[1] https://lwn.net/Articles/456689/